os/bluestore: Move Int64ArrayMergeOperator from BlueStore.cc to bluestore_common.h
Int64ArrayMergeOperator will be used by ZonedFreelistManager as well.
